+++ src/app/_services/index.ts
@@ -12,6 +12,7 @@ export * from './image-group.service'
 export * from './institution.service'
 export * from './log.service'
 export * from './metadata.service'
+export * from './personal-collections.service'
 export * from './slides.service'
 export * from './thumbnail.service'
 export * from './title.service'
