+++ src/app/_services/log.service.ts
@@ -2,7 +2,8 @@ import { HttpClient } from '@angular/common/http'
 import { Injectable } from '@angular/core'
 import { map, take } from 'rxjs/operators'
 
-import { AuthService } from 'app/shared/auth.service'
+// Project Dependencies
+import { AuthService } from '.';
 
 @Injectable()
 export class LogService {
