+++ src/app/app.module.ts
@@ -38,7 +38,7 @@ import { FileUploadModule } from 'ng2-file-upload';
 import { AppComponent } from './app.component'
 import { APP_RESOLVER_PROVIDERS } from './app.resolver'
 import { AppConfig } from './app.service'
-import { Nav, Footer, SearchComponent, PaginationComponent, AssetSearchService, DomUtilityService, ToastComponent, GroupTitleComponent, PromoTooltipComponent } from './shared'
+import { Nav, Footer, SearchComponent, PaginationComponent, ToastComponent, GroupTitleComponent, PromoTooltipComponent } from './shared'
 import { GuideTourComponent } from './shared/tour/tour.component'
 import { NavMenu } from './nav-menu'
 import { AssetFilters } from './asset-filters'
@@ -94,20 +94,28 @@ import { SkyBannerComponent } from './sky-banner/sky-banner.component'
 
 // Application wide providers
 import {
-  AuthService,
+  TypeIdPipe,
+  ScriptService
+} from './shared'
+import { 
+  AccountService,
+  AssetSearchService, 
   AssetService,
+  AuthService,
+  CollectionService, 
+  DomUtilityService,
   FlagService,
   GroupService,
   ImageGroupService,
+  InstitutionService, 
   LogService,
-  TitleService,
-  ToolboxService,
-  TypeIdPipe,
-  ScriptService,
+  MetadataService, 
   PersonalCollectionService,
-  AccountService
-} from './shared'
-import { MetadataService, SlidesService, CollectionService, InstitutionService, ThumbnailService } from './_services'
+  SlidesService, 
+  ThumbnailService, 
+  TitleService,
+  ToolboxService
+} from './_services'
 
 import { LocalPCService } from './_local-pc-asset.service'
 import { AssetFiltersService } from './asset-filters/asset-filters.service'
