+++ src/app/pcollection-page/pcollection-page.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, OnDestroy } from '@angular/core'
-import { HttpClient, HttpHeaders } from '@angular/common/http'
+import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http'
 import { Router, ActivatedRoute } from '@angular/router'
 import { Subscription }   from 'rxjs'
 import { map } from 'rxjs/operators'
@@ -165,7 +165,7 @@ export class PCollectionPage implements OnInit, OnDestroy {
     let statusArray: Array<any> = []
     for (let ssid of this.publishingAssets['ssids']){
       this._assets.getPCImageStatus(ssid).pipe(
-        map(res => {
+        map((res: HttpErrorResponse) => {
           let assetStatus = '' // i.e. available, publishing_que, publishing_failure
           if (res.status){
             assetStatus = 'available'
